backend/modern/requirements.txt
fastapi
pydantic
uvicorn
python-jose
passlib[bcrypt]
pymysql
cryptography
pydantic[email]
pydantic-settings